+++ .idea/workspace.xml
@@ -4,8 +4,10 @@
     <option name="autoReloadType" value="SELECTIVE" />
   </component>
   <component name="ChangeListManager">
-    <list default="true" id="156e1906-a271-4d82-9a2e-9e6ebbc1d4d4" name="Changes" comment="Подключил регулярные выражения.">
+    <list default="true" id="156e1906-a271-4d82-9a2e-9e6ebbc1d4d4" name="Changes" comment="вроде как параметры fp меняются, но отрисовка не происходит">
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/kotlin/drawing/convertation/FractalParameters.kt" beforeDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main/kotlin/guiforfractal/Julia.kt" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/kotlin/guiforfractal/Julia.kt" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/src/main/kotlin/guiforfractal/fileDialogWindow/FileOpenDialogWindow.kt" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/kotlin/guiforfractal/fileDialogWindow/FileOpenDialogWindow.kt" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/src/main/kotlin/guiforfractal/menu.kt" beforeDir="false" afterPath="$PROJECT_DIR$/src/main/kotlin/guiforfractal/menu.kt" afterDir="false" />
     </list>
@@ -423,7 +425,15 @@
       <option name="project" value="LOCAL" />
       <updated>1702330039352</updated>
     </task>
-    <option name="localTasksCounter" value="32" />
+    <task id="LOCAL-00032" summary="вроде как параметры fp меняются, но отрисовка не происходит">
+      <option name="closed" value="true" />
+      <created>1702330987665</created>
+      <option name="number" value="00032" />
+      <option name="presentableId" value="LOCAL-00032" />
+      <option name="project" value="LOCAL" />
+      <updated>1702330987665</updated>
+    </task>
+    <option name="localTasksCounter" value="33" />
     <servers />
   </component>
   <component name="Vcs.Log.Tabs.Properties">
@@ -438,7 +448,6 @@
     </option>
   </component>
   <component name="VcsManagerConfiguration">
-    <MESSAGE value="Написан алгоритм. Не работает, потому что параметры width и height плоскости по умолчанию проинициализированы нулями." />
     <MESSAGE value="Предыдущий коммит содержал ошибки. Этот выполняет сжатие и растяжение фрактала, но при сильном сжатии окна фрактал в окне появляется не полностью, но пропорции &quot;отображаемого&quot; сохраняются" />
     <MESSAGE value="Версия с пары" />
     <MESSAGE value="Версия со второй пары" />
@@ -463,7 +472,8 @@
     <MESSAGE value="Есть ошибка в динамических итерациях" />
     <MESSAGE value="Теперь нет ошибок и сделано 6 номеров" />
     <MESSAGE value="Подключил регулярные выражения." />
-    <option name="LAST_COMMIT_MESSAGE" value="Подключил регулярные выражения." />
+    <MESSAGE value="вроде как параметры fp меняются, но отрисовка не происходит" />
+    <option name="LAST_COMMIT_MESSAGE" value="вроде как параметры fp меняются, но отрисовка не происходит" />
   </component>
   <component name="XSLT-Support.FileAssociations.UIState">
     <expand />
